backend/requirements.txt
Flask==2.3.3
supabase==2.3.1
python-dotenv==1.0.0
Flask-Cors==4.0.0
PyJWT==2.8.0
requests==2.31.0  # Required by supabase
httpx==0.24.1     # Required by supabase 